Enum ecmascript::compiler::value::Value

pub enum Value {
    Undefined,
    Null,
    String(StringRef),
    Object(ObjectRef),
    Boolean(bool),
    I64(i64),
    F64(f64),
    J64(Complex64Ref),
    BigInt(BigIntRef),
    Symbol(SymbolRef),
}

Variants
UndefinedNullString(StringRef)Object(ObjectRef)Boolean(bool)I64(i64)F64(f64)J64(Complex64Ref)BigInt(BigIntRef)Symbol(SymbolRef)
Methods

impl Value

pub const ZERO: Value

pub const ONE: Value

pub const NAN: Value

pub const INFINITY: Value

pub const NEG_INFINITY: Value

pub const NEG_ZERO: Value

pub fn typeof_(&self) -> &'static str

pub fn zero<T: Zero + Into<Value>>() -> T

pub fn one<T: One + Into<Value>>() -> T

pub fn nan<T: Float + Into<Value>>() -> T

pub fn infinity<T: Float + Into<Value>>() -> T

pub fn neg_infinity<T: Float + Into<Value>>() -> T

pub fn neg_zero<T: Float + Into<Value>>() -> T

pub fn is_zero(&self) -> bool

pub fn is_one(&self) -> bool

pub fn is_nan(&self) -> bool

Returns true if this value is NaN and false otherwise.

pub fn is_infinite(&self) -> bool

Returns true if this value is positive infinity or negative infinity and false otherwise.

pub fn is_finite(&self) -> bool

Returns true if this number is neither infinite nor NaN.

pub fn copy(&self) -> Value

pub fn deepcopy(&self) -> Value

pub fn is_undefined(&self) -> bool

pub fn is_null(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_object(&self) -> bool

pub fn is_boolean(&self) -> bool

pub fn is_number(&self) -> bool

pub fn is_int(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_symbol(&self) -> bool

pub fn is_bigint(&self) -> bool

pub fn is_complex(&self) -> bool